Move ethrex-crypto tests into src/tests/ modules
Match the repo test-layout convention (prover/executor src/tests/): split
the inline `mod tests` in lib.rs into per-area files under src/tests/
declared from src/tests/mod.rs.

- src/tests/ecsm_tests.rs       — x-only lincomb2 reconstruction + fallbacks
- src/tests/ecrecover_tests.rs  — full ecsm_ecrecover known-answer/negatives
- src/tests/keccak_tests.rs     — keccak sponge vs reference keccak_hash
