+++ crypto/ethrex-crypto/src/lib.rs
@@ -362,238 +362,4 @@ fn absorb_block(state: &mut [u64; 25], block: &[u8]) {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// secp256k1 curve constant `b = 7`.
-    fn curve_b() -> FieldElement {
-        let mut bytes = [0u8; 32];
-        bytes[31] = 7;
-        FieldElement::from_bytes(&bytes.into()).unwrap()
-    }
-
-    /// Software stand-in for the ECSM precompile: lift `x` to a curve point and
-    /// return `x(k·P)` (parity-invariant, like the real ecall).
-    fn soft_oracle(x: &FieldElement, k: &Scalar) -> Option<FieldElement> {
-        let xn = x.normalize();
-        let y2 = (xn.square() * xn + curve_b()).normalize();
-        let y = Option::<FieldElement>::from(y2.sqrt())?;
-        let p = point_from_xy(&xn, &y.normalize())?;
-        let prod = (p * k).to_affine();
-        Some(affine_xy(&prod)?.0)
-    }
-
-    fn g_times(n: u64) -> ProjectivePoint {
-        ProjectivePoint::GENERATOR * Scalar::from(n)
-    }
-
-    #[test]
-    fn matches_software_lincomb_on_fixed_inputs() {
-        let cases = [
-            (g_times(3), 123_456_789u64, g_times(7), 987_654_321u64),
-            (g_times(11), 2u64.pow(20) + 5, g_times(2), 42u64),
-            (ProjectivePoint::GENERATOR, 7u64, g_times(5), 9u64),
-        ];
-        for (p1, k1, p2, k2) in cases {
-            let (k1, k2) = (Scalar::from(k1), Scalar::from(k2));
-            let expected = ProjectivePoint::lincomb(&p1, &k1, &p2, &k2);
-            let got = lincomb2_with_oracle(&p1, &k1, &p2, &k2, soft_oracle)
-                .expect("non-degenerate inputs must reconstruct");
-            assert_eq!(got.to_affine(), expected.to_affine());
-        }
-    }
-
-    #[test]
-    fn matches_software_lincomb_on_recovery_shape() {
-        // u1·G + u2·R, generator first, like ECDSA recovery.
-        let g = ProjectivePoint::GENERATOR;
-        let r = g_times(0x1234);
-        let u1 = Scalar::from(0xdead_beefu64);
-        let u2 = Scalar::from(0x0bad_f00du64);
-        let expected = ProjectivePoint::lincomb(&g, &u1, &r, &u2);
-        let got = lincomb2_with_oracle(&g, &u1, &r, &u2, soft_oracle)
-            .expect("non-degenerate inputs must reconstruct");
-        assert_eq!(got.to_affine(), expected.to_affine());
-    }
-
-    #[test]
-    fn edge_scalars_fall_back() {
-        let p1 = g_times(3);
-        let p2 = g_times(5);
-        let ok = Scalar::from(12345u64);
-        for bad in [Scalar::ZERO, Scalar::ONE, -Scalar::ONE] {
-            assert!(lincomb2_with_oracle(&p1, &bad, &p2, &ok, soft_oracle).is_none());
-            assert!(lincomb2_with_oracle(&p1, &ok, &p2, &bad, soft_oracle).is_none());
-        }
-    }
-
-    #[test]
-    fn identity_points_fall_back() {
-        let p = g_times(3);
-        let k = Scalar::from(7u64);
-        let id = ProjectivePoint::IDENTITY;
-        assert!(lincomb2_with_oracle(&id, &k, &p, &k, soft_oracle).is_none());
-        assert!(lincomb2_with_oracle(&p, &k, &id, &k, soft_oracle).is_none());
-    }
-
-    #[test]
-    fn cancelling_and_doubling_terms_fall_back() {
-        let p = g_times(3);
-        let k = Scalar::from(7u64);
-        // A = B (doubling chord) and A = −B (Q = O): both share x(A) = x(B).
-        assert!(lincomb2_with_oracle(&p, &k, &p, &k, soft_oracle).is_none());
-        assert!(lincomb2_with_oracle(&p, &k, &(-p), &k, soft_oracle).is_none());
-    }
-
-    // ── Issue 1: known-answer tests for the full ecsm_ecrecover path ─────────
-
-    /// Build a valid ECDSA/secp256k1 signature from (d, kk, msg) using only the
-    /// k256 primitives already imported and return `(sig, recid, expected_addr)`.
-    ///
-    /// `expected_addr` = keccak(X‖Y) of the uncompressed public key, exactly as
-    /// `ecsm_ecrecover` computes it.
-    fn make_ecdsa_fixture(
-        d: Scalar,
-        kk: Scalar,
-        msg: [u8; 32],
-    ) -> ([u8; 64], u8, [u8; 32]) {
-        assert!(!bool::from(d.is_zero()), "private key must be nonzero");
-        assert!(!bool::from(kk.is_zero()), "nonce must be nonzero");
-
-        // Public key Q = d·G.
-        let q = (ProjectivePoint::GENERATOR * d).to_affine();
-        let q_uncompressed = q.to_encoded_point(false);
-        let expected = keccak_hash(&q_uncompressed.as_bytes()[1..65]);
-
-        // R = kk·G; r = reduce(Rx); assert r ≠ 0.
-        let r_point = (ProjectivePoint::GENERATOR * kk).to_affine();
-        let (rx, ry) = affine_xy(&r_point).expect("R is not identity");
-        let r = <Scalar as Reduce<U256>>::reduce_bytes(&rx.to_bytes());
-        assert!(!bool::from(r.is_zero()), "r must be nonzero");
-
-        // recid parity: low bit of Ry (big-endian, byte 31).
-        let recid = ry.normalize().to_bytes()[31] & 1;
-
-        // z = reduce(msg).
-        let z = <Scalar as Reduce<U256>>::reduce_bytes(&FieldBytes::from(msg));
-
-        // s = kk⁻¹ · (z + r·d).
-        let s = kk.invert_vartime().expect("kk is nonzero") * (z + r * d);
-        assert!(!bool::from(s.is_zero()), "s must be nonzero");
-
-        // sig = r (BE, 32 bytes) ‖ s (BE, 32 bytes).
-        let mut sig = [0u8; 64];
-        sig[..32].copy_from_slice(&r.to_bytes());
-        sig[32..].copy_from_slice(&s.to_bytes());
-
-        (sig, recid, expected)
-    }
-
-    #[test]
-    fn ecrecover_known_answer_three_tuples() {
-        // Three distinct (d, kk, msg) tuples — deterministic, no RNG.
-        let tuples: &[(u64, u64, [u8; 32])] = &[
-            (
-                0x0000_0000_0000_0001u64,
-                0x0000_0000_dead_beefu64,
-                {
-                    let mut m = [0u8; 32];
-                    m[31] = 0x42;
-                    m
-                },
-            ),
-            (
-                0x00c0_ffee_dead_beef_u64,
-                0x0123_4567_89ab_cdef_u64,
-                {
-                    let mut m = [0u8; 32];
-                    m[0] = 0xff;
-                    m[31] = 0x01;
-                    m
-                },
-            ),
-            (
-                0x0bad_f00d_1337_cafe,
-                0xfeed_face_0000_0001,
-                {
-                    let mut m = [0u8; 32];
-                    for (i, b) in m.iter_mut().enumerate() {
-                        *b = i as u8;
-                    }
-                    m
-                },
-            ),
-        ];
-
-        for &(d_u64, kk_u64, msg) in tuples {
-            let d = Scalar::from(d_u64);
-            let kk = Scalar::from(kk_u64);
-            let (sig, recid, expected) = make_ecdsa_fixture(d, kk, msg);
-            match ecsm_ecrecover(&sig, recid, &msg) {
-                Ok(got) => assert_eq!(
-                    got, expected,
-                    "ecrecover returned wrong address for d={d_u64:#x} kk={kk_u64:#x}"
-                ),
-                Err(e) => panic!(
-                    "ecrecover failed for d={d_u64:#x} kk={kk_u64:#x}: {e:?}"
-                ),
-            }
-        }
-    }
-
-    #[test]
-    fn ecrecover_rejects_zero_s() {
-        // sig = valid r ‖ 0x00..00 (s = 0) must return InvalidSignature.
-        let mut sig = [0u8; 64];
-        // r = 1 (nonzero, but s = 0 in the second half).
-        sig[31] = 0x01;
-        let msg = [0u8; 32];
-        assert!(
-            matches!(ecsm_ecrecover(&sig, 0, &msg), Err(CryptoError::InvalidSignature)),
-            "expected InvalidSignature for zero s"
-        );
-    }
-
-    #[test]
-    fn ecrecover_rejects_zero_r() {
-        // sig = 0x00..00 ‖ valid s must return InvalidSignature.
-        let mut sig = [0u8; 64];
-        sig[63] = 0x01; // s = 1, r = 0
-        let msg = [0u8; 32];
-        assert!(
-            matches!(ecsm_ecrecover(&sig, 0, &msg), Err(CryptoError::InvalidSignature)),
-            "expected InvalidSignature for zero r"
-        );
-    }
-
-    // ── Issue 2: host-side Keccak sponge tests via injected permutation ───────
-
-    /// Cross-check our sponge body against the trusted `keccak` crate's f1600.
-    fn check_keccak(input: &[u8]) {
-        let got = keccak256_with_permute(input, keccak::f1600);
-        let want = keccak_hash(input);
-        assert_eq!(
-            got, want,
-            "keccak256 mismatch for {}-byte input",
-            input.len()
-        );
-    }
-
-    #[test]
-    fn keccak_sponge_matches_trusted_permutation() {
-        // Empty input.
-        check_keccak(&[]);
-        // One byte.
-        check_keccak(&[0xab]);
-        // 135 bytes — RATE-1: padding lands on byte 135 (0x01) and byte 135 is
-        // also the last byte (0x80), so both bits land on the same byte: 0x81.
-        check_keccak(&[0x5a; 135]);
-        // Exactly RATE (136): fills one full block, final block is all-padding.
-        check_keccak(&[0x3c; 136]);
-        // RATE+1: one full block + one-byte remainder.
-        check_keccak(&[0x7e; 137]);
-        // Multi-block: ~1.5 × RATE (200 bytes), deterministic pattern.
-        let long: Vec<u8> = (0u8..200).collect();
-        check_keccak(&long);
-    }
-}
+mod tests;
