+++ crypto/ethrex-crypto/src/tests/mod.rs
@@ -1,0 +1,6 @@
+#[cfg(test)]
+pub mod ecrecover_tests;
+#[cfg(test)]
+pub mod ecsm_tests;
+#[cfg(test)]
+pub mod keccak_tests;
